+++ R/as_pal.R
@@ -24,7 +24,7 @@
 #'
 #' @examples
 #'   # create LUV palette-function
-#'   pal_blues <- palette_bezier(mat_luv_blues)
+#'   pal_blues <- pal_luv_bezier(mat_luv_blues)
 #'
 #'   # create hex palette-function
 #'   pal_blues_hex <- as_pal_hex(pal_blues)