+++ R/rescale_pal_luv.R
@@ -8,7 +8,7 @@
 #' @return `function` with S3 class `"cpath_pal_luv"`
 #' @examples
 #'   # create original palette
-#'   pal_blues <- palette_bezier(mat_luv_blues)
+#'   pal_blues <- pal_luv_bezier(mat_luv_blues)
 #'
 #'   # create rescaler
 #'   rsc_reverse <- rescaler_x(c(1, 0))