+++ tests/testthat/test-as_pal.R
@@ -1,4 +1,4 @@
-pal_blues <- palette_bezier(mat_luv_blues)
+pal_blues <- pal_luv_bezier(mat_luv_blues)
 
 test_that("as_pal_hex works", {